Cleanup shuffle_mol
