+++ tests/test_spatial.py
@@ -1,5 +1,4 @@
 import os
-from random import shuffle
 
 from nose.tools import assert_equal
 from sklearn.utils.testing import (assert_almost_equal,
@@ -9,24 +8,12 @@ from sklearn.utils.testing import (assert_almost_equal,
 import numpy as np
 
 import oddt
-from oddt.utils import is_openbabel_molecule
 from oddt.spatial import (angle,
                           dihedral,
                           rmsd,
                           distance,
                           rotate)
-
-
-def shuffle_mol(mol):
-    """Randomly reorder molecule atoms and return a shuffled copy of input."""
-    new_mol = mol.clone
-    new_order = list(range(len(mol.atoms)))
-    shuffle(new_order)
-    if is_openbabel_molecule(mol):
-        new_mol.OBMol.RenumberAtoms([i + 1 for i in new_order])
-    else:
-        new_mol.Mol = oddt.toolkits.rdk.Chem.RenumberAtoms(new_mol.Mol, new_order)
-    return new_mol
+from .utils import shuffle_mol
 
 
 test_data_dir = os.path.dirname(os.path.abspath(__file__))
